+++ cpp/src/io/utilities/data_casting.cu
@@ -466,7 +466,7 @@ CUDF_KERNEL void parse_fn_string_parallel(str_tuple_it str_tuples,
       // FIXME(HIP/AMD): work-around for SWDEV-470886 through explicit initialization
       size_type istring = 0;
       if (lane == 0) { istring = atomicAdd(str_counter, 1); }
-      return __shfl_sync(LANE_MASK_ALL, istring, 0);
+      return __shfl_sync((uint64_t) LANE_MASK_ALL, istring, 0);
     } else {
       // Ensure lane 0 doesn't update istring before all threads have read the previous iteration's
       // istring value
@@ -598,7 +598,7 @@ CUDF_KERNEL void parse_fn_string_parallel(str_tuple_it str_tuples,
         __shared__ typename SlashScan::TempStorage temp_slash[num_warps];
         SlashScan(temp_slash[warp_id]).InclusiveScan(curr, scanned, composite_op);
         is_escaping_backslash = scanned.get(init_state);
-        init_state            = __shfl_sync(LANE_MASK_ALL, is_escaping_backslash, BLOCK_SIZE - 1);
+        init_state            = __shfl_sync((uint64_t) LANE_MASK_ALL, is_escaping_backslash, BLOCK_SIZE - 1);
         __syncwarp();
         is_slash.shift(warp_id);
         is_slash.set_bits(warp_id, is_escaping_backslash);
@@ -631,7 +631,7 @@ CUDF_KERNEL void parse_fn_string_parallel(str_tuple_it str_tuples,
       }
       // Make sure all threads have no errors before continuing
       if constexpr (is_warp) {
-        error = __any_sync(LANE_MASK_ALL, error);
+        error = __any_sync((uint64_t) LANE_MASK_ALL, error);
       } else {
         using ErrorReduce = hipcub::BlockReduce<bool, BLOCK_SIZE>;
         __shared__ typename ErrorReduce::TempStorage temp_storage_error;
@@ -751,7 +751,7 @@ CUDF_KERNEL void parse_fn_string_parallel(str_tuple_it str_tuples,
         }
         offset += this_num_out;
         if constexpr (is_warp) {
-          last_offset = __shfl_sync(LANE_MASK_ALL, offset, BLOCK_SIZE - 1);
+          last_offset = __shfl_sync((uint64_t) LANE_MASK_ALL, offset, BLOCK_SIZE - 1);
         } else {
           __syncthreads();
           if (threadIdx.x == BLOCK_SIZE - 1) last_offset = offset;
