+++ cpp/src/json/json_path.cu
@@ -953,7 +953,7 @@ __launch_bounds__(block_size) CUDF_KERNEL
 
   size_type warp_valid_count{0};
 
-  bitmask_type active_threads = __ballot_sync(LANE_MASK_ALL, tid < col.size());
+  bitmask_type active_threads = __ballot_sync((uint64_t) LANE_MASK_ALL, tid < col.size());
   while (tid < col.size()) {
     bool is_valid         = false;
     string_view const str = col.element<string_view>(tid);
@@ -977,7 +977,7 @@ __launch_bounds__(block_size) CUDF_KERNEL
 
     // validity filled in only during the output step
     if (out_validity.has_value()) {
-      bitmask_type mask = __ballot_sync(active_threads, is_valid);
+      bitmask_type mask = __ballot_sync((uint64_t) active_threads, is_valid);
       // 0th lane of the warp writes the validity
       if (!(tid % cudf::detail::warp_size)) {
         THRUST_OPTIONAL_VALUE(out_validity)[cudf::word_index(tid)] = mask;
@@ -986,7 +986,7 @@ __launch_bounds__(block_size) CUDF_KERNEL
     }
 
     tid += stride;
-    active_threads = __ballot_sync(active_threads, tid < col.size());
+    active_threads = __ballot_sync((uint64_t) active_threads, tid < col.size());
   }
 
   // sum the valid counts across the whole block
